requirements.txt
timm
opencv-python
tqdm
torchstat
torchsummary
Pillow
h5py
thop
imageio
tensorboardX
scikit-learn
efficientnet-pytorch

